Roll src/third_party/ffmpeg/ 7e1d53a09..cf7ee6598 (1000 commits)
https://chromium.googlesource.com/chromium/third_party/ffmpeg.git/+log/7e1d53a09f9a..cf7ee6598c6f

$ git log 7e1d53a09..cf7ee6598 --date=short --no-merges --format='%ad %ae %s'
2021-06-14 tmathmeyer Remove avcodec_decode_video2 from sigs
2021-06-04 tmathmeyer Initialize variables to fix compilation warning
2021-06-02 tmathmeyer Cherry pick change from ffmpeg::master onto sushi branch for roll
2021-05-15 tmathmeyer Update FFMpeg
2021-05-07 ffmpeg doc/muxers: note movie_timescale option in movenc
(...)
2021-03-10 andreas.rheinhardt dnn/dnn_backend_native: Fix typo in log message
2021-03-10 andreas.rheinhardt dnn/dnn_backend_native: Don't use asserts for checks
2021-03-10 andreas.rheinhardt dnn/dnn_backend_native: Fix leak in case parsing options fails
2021-03-10 andreas.rheinhardt dnn/dnn_backend_native: Avoid allocation for checking file magic
2021-03-10 jb Doc: Tech Resolution Process

Created with:
  roll-dep src/third_party/ffmpeg

Bug: 1208571
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2965899
Commit-Queue: Ted Meyer <tmathmeyer@chromium.org>
Cr-Commit-Position: refs/heads/master@{#893698}

+++ media/filters/audio_decoder_unittest.cc
@@ -109,7 +109,7 @@ void SetDiscardPadding(AVPacket* packet,
   }
 
   // If the timestamp is positive, try to use FFmpeg's discard data.
-  int skip_samples_size = 0;
+  size_t skip_samples_size = 0;
   const uint32_t* skip_samples_ptr =
       reinterpret_cast<const uint32_t*>(av_packet_get_side_data(
           packet, AV_PKT_DATA_SKIP_SAMPLES, &skip_samples_size));
